#advanced

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,7 +7,7 @@
 
 int main(void)
 {
-	unsigned long int i, nb1, nb2, aux;
+	unsigned long long int i, nb1, nb2, aux;
 
 	nb1 = 0;
 	nb2 = 1;
